Montei os sql para upgrade de v2 to v3 e fiz ajustes em alguns tipos

+++ src/Request/Configuration/Update/v2tov3/create_procedures.sql
@@ -17,8 +17,8 @@ BEGIN
   SET v_upper_is_part_of = CONCAT(UPPER(LEFT(table_is_part_of, 1)),SUBSTRING(table_is_part_of, 2));
 
   SET @v_sql_text = CONCAT(
-    'INSERT INTO `thing_has_thing` (`idHasPart`, `typeHasPart`, `idIsPartOf`, `typeIsPartOf`, `caption`, `position`, `representativeOfPage`)
-      SELECT ', table_has_part, '.`thing`,
+          'INSERT INTO `thing_has_thing` (`idHasPart`, `typeHasPart`, `idIsPartOf`, `typeIsPartOf`, `caption`, `position`, `representativeOfPage`)
+            SELECT ', table_has_part, '.`thing`,
       ''', v_upper_has_part, ''',
       ', table_is_part_of, '.`thing`,
       ''', v_upper_is_part_of, ''',
@@ -28,7 +28,7 @@ BEGIN
       FROM ', v_table_has, '
       JOIN `', table_has_part, '` ON ', v_table_has, '.`', v_id_name_has_part, '` = `', table_has_part, '`.`', v_id_name_has_part, '`
       JOIN `', table_is_part_of, '` ON ', v_table_has, '.`', v_id_name_is_part_of, '` = `', table_is_part_of, '`.`', v_id_name_is_part_of, '`'
-  );
+                    );
 
   PREPARE stmt FROM @v_sql_text;
   EXECUTE stmt;
@@ -39,21 +39,20 @@ CREATE PROCEDURE set_image_in_thing(IN table_name VARCHAR(64))
 BEGIN
   DECLARE v_table_has_part VARCHAR(128);
   DECLARE v_id_name VARCHAR(128);
-  DECLARE v_sql_text TEXT;
   -- Monta nomes dinâmicos
   SET v_table_has_part = CONCAT(table_name, '_has_imageObject');
   SET v_id_name = CONCAT('id', table_name);
   -- Monta SQL dinâmico
-  SET v_sql_text = CONCAT(
-    'UPDATE `thing`
-      JOIN `', table_name, '` AS tb1 ON tb1.thing = thing.idthing
+  SET @v_sql_text = CONCAT(
+          'UPDATE `thing`
+            JOIN `', table_name, '` AS tb1 ON tb1.thing = thing.idthing
       JOIN `', v_table_has_part, '` AS tb_has ON tb_has.', v_id_name, ' = tb1.', v_id_name, '
       JOIN imageObject ON imageObject.idimageObject = tb_has.idimageObject
       JOIN mediaObject ON mediaObject.idmediaObject = imageObject.mediaObject
      SET `thing`.image = mediaObject.contentUrl'
-                    );
+                   );
   -- Executa o SQL dinâmico
-  PREPARE stmt FROM v_sql_text;
+  PREPARE stmt FROM @v_sql_text;
   EXECUTE stmt;
   DEALLOCATE PREPARE stmt;
 END;
@@ -75,7 +74,7 @@ BEGIN
   DECLARE cur2 CURSOR FOR
     SELECT table_name, index_name
     FROM information_schema.statistics
-    WHERE table_schema = schema_name AND index_name != 'PRIMARY' AND table_name != 'user' AND table_name != 'user_history' AND table_name != 'user_privileges';
+    WHERE table_schema = schema_name AND index_name <> 'PRIMARY' AND table_name <> 'user';
 
   DECLARE CONTINUE HANDLER FOR NOT FOUND SET done = 1;
 
@@ -102,13 +101,17 @@ BEGIN
     IF done THEN
       LEAVE read_loop2;
     END IF;
-    SET @drop_idx = CONCAT('ALTER TABLE `', schema_name, '`.`', tablename, '` DROP INDEX `', indexname, '`');
-    PREPARE stmt2 FROM @drop_idx;
-    EXECUTE stmt2;
-    DEALLOCATE PREPARE stmt2;
+    IF EXISTS(
+      SELECT 1 FROM information_schema.statistics
+      WHERE table_schema = schema_name
+        AND table_name = tablename
+        AND index_name = indexname
+    ) THEN
+      SET @drop_idx = CONCAT('ALTER TABLE `', schema_name, '`.`', tablename, '` DROP INDEX `', indexname, '`');
+      PREPARE stmt2 FROM @drop_idx;
+      EXECUTE stmt2;
+      DEALLOCATE PREPARE stmt2;
+    END IF;
   END LOOP;
   CLOSE cur2;
 END;
-
-
-call drop_keys('pirenopolis');
